Move two tests to contributor app
