+++ evap/staff/tests.py
@@ -966,82 +966,6 @@ class ContributionFormsetTests(TestCase):
         formset = ContributionFormset(instance=course, data=data)
         self.assertTrue(formset.is_valid())
 
-    def test_editors_cannot_change_responsible(self):
-        """
-            Asserts that editors cannot change the responsible of a course
-            through POST-hacking. Regression test for #504.
-        """
-        course = mommy.make(Course)
-        user1 = mommy.make(UserProfile)
-        user2 = mommy.make(UserProfile)
-        questionnaire = mommy.make(Questionnaire, is_for_contributors=True)
-        contribution1 = mommy.make(Contribution, course=course, contributor=user1, responsible=True, can_edit=True, comment_visibility=Contribution.ALL_COMMENTS, questionnaires=[questionnaire])
-
-        EditorContributionFormset = inlineformset_factory(Course, Contribution, formset=ContributionFormSet, form=EditorContributionForm, extra=0, exclude=('course',))
-
-        data = {
-            'contributions-TOTAL_FORMS': 1,
-            'contributions-INITIAL_FORMS': 1,
-            'contributions-MAX_NUM_FORMS': 5,
-            'contributions-0-id': contribution1.pk,
-            'contributions-0-course': course.pk,
-            'contributions-0-questionnaires': [questionnaire.pk],
-            'contributions-0-order': 1,
-            'contributions-0-responsibility': "RESPONSIBLE",
-            'contributions-0-comment_visibility': "ALL",
-            'contributions-0-contributor': user1.pk,
-        }
-
-        formset = EditorContributionFormset(instance=course, data=data.copy())
-        self.assertTrue(formset.is_valid())
-
-        self.assertTrue(course.contributions.get(responsible=True).contributor == user1)
-        data["contributions-0-contributor"] = user2.pk
-        formset = EditorContributionFormset(instance=course, data=data.copy())
-        self.assertTrue(formset.is_valid())
-        formset.save()
-        self.assertTrue(course.contributions.get(responsible=True).contributor == user1)
-
-
-class ContributionFormsetWebTests(WebTest):
-    csrf_checks = False
-
-    def test_form_ordering(self):
-        """
-            Asserts that the contribution formset is correctly sorted,
-            and that an ordering changed by the user survives the reload
-            when the user submits the form with errors.
-            Regression test for #456.
-        """
-        course = mommy.make(Course, pk=1, state="prepared")
-        user1 = mommy.make(UserProfile)
-        user2 = mommy.make(UserProfile)
-        questionnaire = mommy.make(Questionnaire, is_for_contributors=True)
-        contribution1 = mommy.make(Contribution, course=course, contributor=user1, responsible=True, can_edit=True, comment_visibility=Contribution.ALL_COMMENTS, questionnaires=[questionnaire], order=1)
-        contribution2 = mommy.make(Contribution, course=course, contributor=user2, responsible=False, can_edit=True, comment_visibility=Contribution.ALL_COMMENTS, questionnaires=[questionnaire], order=2)
-
-        # almost everything is missing in this set of data,
-        # so we're guaranteed to have some errors
-        data = {
-            "contributions-TOTAL_FORMS": 2,
-            "contributions-INITIAL_FORMS": 2,
-            "contributions-MIN_NUM_FORMS": 0,
-            "contributions-MAX_NUM_FORMS": 1000,
-            "contributions-0-id": contribution1.id,
-            "contributions-1-id": contribution2.id,
-            "operation": "save"
-        }
-
-        data["contributions-0-order"] = 1
-        data["contributions-1-order"] = 2
-        response = str(self.app.post("/contributor/course/1/edit", data, user=user1))
-        self.assertTrue(response.index("id_contributions-1-id") > response.index("id_contributions-0-id"))
-
-        data["contributions-0-order"] = 2
-        data["contributions-1-order"] = 1
-        response = str(self.app.post("/contributor/course/1/edit", data, user=user1))
-        self.assertFalse(response.index("id_contributions-1-id") > response.index("id_contributions-0-id"))
-
 
 class ArchivingTests(WebTest):
     fixtures = ['minimal_test_data']
